feat(feed): triangle = new questions only, convergence on Home (+ build fix)
Completes the two-tier home stream and fixes the Vercel build it tripped.

Two-tier logic:
- build-stream: the milestone "triangle" row now drops any question the
  viewer has already attempted on ANY surface (priorById carries every
  prior answer, right or wrong), so it carries ONLY new/unanswered
  questions. A milestone that empties out is suppressed rather than shown
  as a contentless streak card. The both-correct overlap is told by the
  convergence row; the viewer's own misses live in Catch-Up.
- activity-stream: the convergence card ("you and X both knew these") is
  now homeEligible, so it surfaces on Home alongside the triangle rows
  instead of only in Lately.

Build fix (turbopack "Module not found: dns/fs/net/tls" on this branch):
activity-stream is the client-shared, DB-free transform, but it imported
HOME_TOP3_ELIGIBLE_TYPES (a value) from write-activity, which imports the
pg DB layer — dragging pg into the browser bundle via FeedList. Extracted
the activity-type vocabulary + home-eligible set into a pure
@/lib/activity-types module; write-activity re-exports them for server
callers, and activity-stream imports straight from the pure module.
